restructures in to project / package / modules
